Orders now working
The delivery crew should now be assigned properly when the manager edits the orders. Also, the manager and delivery crew roles should be assigned to the users correctly only if they are not yet parts of the group. (The same goes for removing the user from those groups if they are already part of the groups.)